+++ src/person/person.service.ts
@@ -1,9 +1,10 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, HttpStatus } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Person } from './person.entity';
 import { Media } from '../media/media.entity';
 import { CreatePersonDto } from './person.dto';
+import { BusinessException } from '../common/exceptions/business.exception';
 
 @Injectable()
 export class PersonService {
@@ -29,6 +30,19 @@ export class PersonService {
     const person = new Person();
     Object.assign(person, createPersonDto);
 
+    if (createPersonDto.icon) {
+      const media = await this.mediaRepository.findOneBy({
+        id: createPersonDto.icon,
+      });
+      if (!media) {
+        throw new BusinessException(
+          `Media with ID ${createPersonDto.icon} not found`,
+          HttpStatus.NOT_FOUND,
+        );
+      }
+      person.icon = media;
+    }
+
     return this.personRepository.save(person);
   }
 
@@ -36,9 +50,15 @@ export class PersonService {
     id: number,
     updatePersonDto: CreatePersonDto,
   ): Promise<Person | null> {
-    const person = await this.personRepository.findOneBy({ id });
+    const person = await this.personRepository.findOne({
+      where: { id },
+      relations: ['icon'],
+    });
     if (!person) {
-      throw new Error('Person not found');
+      throw new BusinessException(
+        `Person with ID ${id} not found`,
+        HttpStatus.NOT_FOUND,
+      );
     }
 
     Object.assign(person, updatePersonDto);
@@ -48,12 +68,15 @@ export class PersonService {
         id: updatePersonDto.icon,
       });
       if (!media) {
-        throw new Error('Media not found');
+        throw new BusinessException(
+          `Media with ID ${updatePersonDto.icon} not found`,
+          HttpStatus.NOT_FOUND,
+        );
       }
       person.icon = media;
     }
 
-    await this.personRepository.update(id, person);
+    await this.personRepository.save(person);
     return this.personRepository.findOne({
       where: { id },
       relations: ['icon'],
@@ -61,6 +84,13 @@ export class PersonService {
   }
 
   async remove(id: number): Promise<void> {
+    const person = await this.personRepository.findOneBy({ id });
+    if (!person) {
+      throw new BusinessException(
+        `Person with ID ${id} not found`,
+        HttpStatus.NOT_FOUND,
+      );
+    }
     await this.personRepository.delete(id);
   }
 }
